requirements.txt

Flask==2.2.5
PyPDF2==3.0.1
google-cloud-aiplatform==1.26.0
